Add log module with a lot of smelly code
This code smells, it's ugly and I have yet to find out why.
While the new function solves the repeating aspect of creating
a new Info, Debug, and Error logger, returning an array of
os.File pointers is just ugly. Which needs to be closed when a
log.Fatal triggers.

Definetly needs refactoring.